+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - **Controls.Global**
   - Improved value comparison (static method `$isEqual`) to compare by `.id`. So now don't need to worry about complex objects with **id's** (mostly related to `SelectControl.$options.items` & `RadioControl.$options.items`)
   - Added readonly property `$isRequired` based on `$options.validations.required`
+- [CalendarControl](src/controls/calendar.ts). [DateControl](src/controls/date.ts). `$options.startWith` can be string-date
 
 ## 0.6.1 (Jun 26, 2023)